requirements.txt
tqdm
numpy
beartype
biopython
einops
MEGABYTE-pytorch==0.2.1
torch==2.3.0